pyproject.toml
[project]
name = "zen-portfolio-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "LJ Li",email = "liliangjya@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "httpx[socks] (>=0.28.1,<0.29.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "tenacity (>=9.0.0,<10.0.0)",
    "apscheduler (>=3.10.1,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "python-redis-lock (>=4.0.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.25.3"

